refactor(orchestrator): extract workspace ops into orchestration/workspace-ops.ts (seam 24, #1526)
Behavior-preserving move of gatherTodayFacts, previewMemoryActionEvent,
maybeRunFileHygiene, autoConsolidateIdentity, getWearablesService,
validateLocalLlmModel, runPatternReinforcement(+Fanout), and
flushAccessTracking (686 LOC, 20 deps, 2 written fields) into
WorkspaceOpsCoordinator (late-binding deps rule, seams 18-23).
IDENTITY_CONSOLIDATE_THRESHOLD static widened from private for the
coordinator read; day-summary helpers imported from ../orchestrator.js.

orchestrator.ts: 6,476 -> 5,886 LOC.
